third_party/blink/web_tests/virtual/stable/webexposed/null-attributes-shared-worker-expected.txt

This test interface attributes which should be null in SharedWorkerGlobalScope.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: resources/null-attributes-shared-worker.js
FAIL [Worker] navigator.serial should be null (of type object). Was undefined (of type undefined).
PASS [Worker] navigator.usb is null
FAIL [Worker] navigator.wakeLock should be null (of type object). Was undefined (of type undefined).
PASS successfullyParsed is true

TEST COMPLETE
